xtool/internal/clangtool:split with libclang
